Fixing warnings with MPI configurations.
Former-commit-id: 93fe9a63065254af557cc4db8553d5eb4778292f

+++ plugins/examples/Tau_plugin_example_disable_instrumentation_runtime.cpp
@@ -112,7 +112,7 @@ int processInstrumentationRequests(char *fname)
          * "#foo" becomes #foo and is passed on to the exclude list. */
         if (inbuf[0] == '"') {
           char *exclude = strdup(&inbuf[1]);
-          for (int i = 0; i < strlen(exclude); i++) {
+          for (size_t i = 0; i < strlen(exclude); i++) {
             if (exclude[i] == '"') {
               exclude[i] = '\0';
               break;
@@ -143,7 +143,7 @@ int processInstrumentationRequests(char *fname)
          * "#foo" becomes #foo and is passed on to the exclude list. */
         if (inbuf[0] == '"') {
           char *exclude = strdup(&inbuf[1]);
-          for (int i = 0; i < strlen(exclude); i++) {
+          for (size_t i = 0; i < strlen(exclude); i++) {
             if (exclude[i] == '"') {
               exclude[i] = '\0';
               break;
